mcp-service/requirements.txt
aiomysql>=0.1.1
annotated-types==0.7.0
anyio==4.10.0
asgi-correlation-id==4.3.4
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langchain-core==0.3.74
langchain-openai==0.3.30
langsmith==0.4.14
loguru==0.7.3
markdown-it-py==4.0.0
fastmcp>=2.3.2
mdurl==0.1.2
openai==1.100.1
orjson==3.11.2
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.33
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
zstandard==0.24.0
pytz == 2025.2
